Refactor lazy loading utilities into separate utils file
